change package zmx/server to zmx/diagnostics
