+++ src/main/scala/zio/zmx/diagnostics/ZMXError.scala
@@ -1,4 +1,4 @@
-package zio.zmx.server
+package zio.zmx.diagnostics
 
 sealed trait ZMXError
 case class UnknownZMXCommand(command: String) extends ZMXError
