+++ src/main/scala/zio/zmx/diagnostics/ZMXProtocol.scala
@@ -14,9 +14,28 @@
  * limitations under the License.
  */
 
-package zio.zmx.server
+package zio.zmx.diagnostics
 
-final case class ZMXServerRequest(
-  command: String,
-  args: Option[List[String]]
-)
+object ZMXProtocol {
+
+  sealed trait Command
+
+  object Command {
+    case object Test      extends Command
+    case object Stop      extends Command
+    case object FiberDump extends Command
+  }
+
+  sealed trait Message
+
+  object Message  {
+    final case class Simple(message: String) extends Message {
+      override def toString: String = message
+    }
+  
+    final case class FiberDump(dumps: List[String]) extends Message {
+      override def toString: String = dumps.mkString("\n")
+    }
+  }
+  
+}
