+++ src/main/scala/zio/zmx/diagnostics/parser/package.scala
@@ -1,4 +1,4 @@
-package zio.zmx.server
+package zio.zmx.diagnostics
 
 import zio._
