+++ src/test/scala/zio/zmx/diagnostics/ZMXClientSpec.scala
@@ -14,16 +14,15 @@
  * limitations under the License.
  */
 
-package zio.zmx.server.parser
+package zio.zmx.diagnostics
 
-import zio.zmx.server._
 import zio.test.Assertion._
 import zio.test._
 
-object RespZMXParserSpec extends DefaultRunnableSpec {
+object ZMXClientSpec extends DefaultRunnableSpec {
   def spec =
-    suite("RespZMXParserSpec")(
-      suite("Using the RESP ZMX parser")(
+    suite("ZMXClientSpec")(
+      suite("Using the ZMXClient")(
         test("zmx test generating a successful command") {
           val p: String = ZMXClient.generateRespCommand(args = List("foobar"))
           assert(p)(equalTo("*1\r\n$6\r\nfoobar\r\n"))
@@ -35,22 +34,7 @@ object RespZMXParserSpec extends DefaultRunnableSpec {
         test("zmx test generating a successful empty command") {
           val p: String = ZMXClient.generateRespCommand(args = List())
           assert(p)(equalTo("*0\r\n"))
-        },
-        test("zmx test generating a simple reply") {
-          val value = RespZMXParser.asString(ZMXProtocol.Message.Simple("foobar"), Success)
-          assert(value)(equalTo("+foobar"))
-
-        },
-        test("zmx test generating a fiber dump list reply") {
-          val d                  = List("foo", "bar", "baz")
-          val value = RespZMXParser.asString(ZMXProtocol.Message.FiberDump(d), Success)
-          assert(value)(equalTo("*3\r\n+foo\r\n+bar\r\n+baz\r\n"))
-        },
-        test("zmx test generating a fail reply") {
-          val value = RespZMXParser.asString(ZMXProtocol.Message.Simple("foobar"), Fail)
-          assert(value)(equalTo("-foobar"))
         }
       )
-      //TODO add fromString tests
     )
 }
